Fixup previous commit about enableSound
- one bool too many in RosegardenDocument
- don't initialize sound when SequenceManager is created
    (I need SequenceManager because it has the fastforward() code;
     the alternative would be to move that code elsewhere, but I don't know where)

git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@14640 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ src/gui/seqmanager/SequenceManager.cpp
@@ -120,9 +120,12 @@ SequenceManager::~SequenceManager()
 
     SEQMAN_DEBUG << "SequenceManager::~SequenceManager()";
     delete m_compositionMapper;
-    m_metronomeMapper->removeOwner();
-    m_tempoSegmentMapper->removeOwner();
-    m_timeSigSegmentMapper->removeOwner();
+    if (m_metronomeMapper)
+        m_metronomeMapper->removeOwner();
+    if (m_tempoSegmentMapper)
+        m_tempoSegmentMapper->removeOwner();
+    if (m_timeSigSegmentMapper)
+        m_timeSigSegmentMapper->removeOwner();
 }
 
 void
@@ -170,9 +173,10 @@ SequenceManager::setDocument(RosegardenDocument *doc, QWidget *parentWidget)
     connect(CommandHistory::getInstance(), SIGNAL(commandExecuted()),
             this, SLOT(update()));
 
-    resetCompositionMapper();
-    
-    populateCompositionMapper();
+    if (doc->isSoundEnabled()) {
+        resetCompositionMapper();
+        populateCompositionMapper();
+    }
 }
 
 void SequenceManager::setTrackEditor(TrackEditor *trackEditor)
